+++ src/components/Common/Header/index.jsx
@@ -2,36 +2,26 @@ import { Link, useLocation, useNavigate } from 'react-router-dom';
 import { useMemo, useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 import useMediaQuery from '@mui/material/useMediaQuery';
-import {
-  AppBar,
-  Box,
-  Container,
-  Menu,
-  MenuItem,
-  ListItemIcon,
-  ListItemText,
-} from '@mui/material';
+import { AppBar, Box, Container } from '@mui/material';
 import HeaderTabs from './HeaderTabs.jsx';
 import HeaderFab from './HeaderFab.jsx';
 import Logo from 'components/Common/Logo.jsx';
 import { useAuthState } from 'contexts/authContext.jsx';
 import { useCustomThemeDispatch } from 'contexts/CustomThemeProvider.jsx';
-import Brightness7Icon from '@mui/icons-material/Brightness7';
-import Brightness4Icon from '@mui/icons-material/Brightness4';
+
 import { useTheme } from '@emotion/react';
-import LogoutIcon from '@mui/icons-material/Logout';
+
 import { useAuthDispatch } from 'contexts/authContext.jsx';
+import HeaderDrawerMenu from './HeaderDrawerMenu.jsx';
 
 const Header = () => {
   const location = useLocation();
-  const navigate = useNavigate();
   const { isLoggedIn } = useAuthState();
-  const { onLogout } = useAuthDispatch();
   const isMainPage = useMemo(() => location.pathname === '/', [location]);
   const isLoginPage = useMemo(() => location.pathname === '/login', [location]);
   const theme = useTheme();
   const mdDownMatches = useMediaQuery(theme.breakpoints.down('md'));
-  const { onToggleTheme } = useCustomThemeDispatch();
+  const navigate = useNavigate();
 
   const [anchorEl, setAnchorEl] = useState(null);
   const open = useMemo(() => !!anchorEl, [anchorEl]);
@@ -43,21 +33,6 @@ const Header = () => {
   const handleCloseMenu = useCallback(() => {
     setAnchorEl(null);
   }, []);
-
-  const handleNavigateMyPage = useCallback(() => {
-    navigate && navigate(`/user`);
-    handleCloseMenu();
-  }, [navigate, handleCloseMenu]);
-
-  const handleToggleTheme = useCallback(() => {
-    onToggleTheme && onToggleTheme();
-  }, [onToggleTheme]);
-
-  const handleLogout = useCallback(() => {
-    onLogout && onLogout();
-    handleCloseMenu();
-    navigate('/');
-  }, [onLogout, handleCloseMenu, navigate]);
 
   const handleNavigateLoginPage = useCallback(() => {
     navigate('/login');
@@ -126,62 +101,12 @@ const Header = () => {
                 onClickMenu={handleOpenMenu}
                 onClickLoginButton={handleNavigateLoginPage}
               />
-              <Menu
-                id="menu"
-                anchorEl={anchorEl}
+              <HeaderDrawerMenu
                 open={open}
-                onClose={handleCloseMenu}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'right',
-                }}
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-              >
-                <MenuItem
-                  onClick={handleNavigateMyPage}
-                  sx={{
-                    background: 'transparent',
-                  }}
-                >
-                  <ListItemIcon />
-                  <ListItemText>마이 페이지</ListItemText>
-                </MenuItem>
-
-                <MenuItem
-                  onClick={handleToggleTheme}
-                  sx={{
-                    background: 'transparent',
-                  }}
-                >
-                  <ListItemIcon>
-                    {theme.palette.mode === 'dark' ? (
-                      <Brightness7Icon />
-                    ) : (
-                      <Brightness4Icon />
-                    )}
-                  </ListItemIcon>
-                  <ListItemText>
-                    {theme.palette.mode === 'dark'
-                      ? '밝은 테마'
-                      : '어두운 테마'}
-                  </ListItemText>
-                </MenuItem>
-
-                <MenuItem
-                  onClick={handleLogout}
-                  sx={{
-                    background: 'transparent',
-                  }}
-                >
-                  <ListItemIcon>
-                    <LogoutIcon />
-                  </ListItemIcon>
-                  <ListItemText>로그아웃</ListItemText>
-                </MenuItem>
-              </Menu>
+                anchorEl={anchorEl}
+                theme={theme}
+                onCloseMenu={handleCloseMenu}
+              />
             </Box>
           </Box>
         </Container>
